Add buttons and DOM modifications
The rock paper scissors game now has buttons to interact with instead
of needing to use the console prompt. The playRound function was
modified to handle the entirety of the text of the game result instead
of splitting it between game() and playRound(). This will keep functions
more clear. The game will also display the results in a div container.