Fix scoring and styling
The styling of the sheet was now made to have a header, footer, and
the center content will stretch across the rest of the given space
using flexbox. An error in the scoring would happen if the player
reached 5 points regardless of it was less than the computer's score.
Now the code takes into account who has more points on top of reaching
five points.